Ensure deleted events are removed
To retrieve deleted events, we have to use an incremental sync
method: https://developers.google.com/calendar/v3/sync

Incremental sync requires a syncToken, which is only retrieved on
the last page of results. This means we cannot limit the number of
pages. In addition, there are a number of query parameters that
cannot be specified together with a syncToken, including `orderBy`
and `timeMin` (or `timeMax`).

Page limits and query parameters have been removed so that the
first time a sync takes place, all events are imported. Then, on
subsequent syncs, only events which have changed are returned
(including deleted events).