Prevent queue from growing when an import error occurs
If an unexpected error occurs during the import of a queued
calendar, it will persist in the queue to be retried later. This
can result in queue items which are never deleted.

This causes duplicate queue items to pile up, because each time
cron runs, a new item is added to the queue and never deleted.
Eventually the queue can reach a state where it's essentially
trying to import the same calendar thousands of times, incremented
by n+1 each time cron runs.

This approach deletes the entire queue every time cron runs to
ensure queue items never pile up, even if a specific calendar
encounters an error each time it's imported.